Move high-level API to index file
